backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2
python-docx>=1.1.0
pydantic>=2.5.3
python-multipart>=0.0.6
aiofiles>=23.2.1
chromadb>=0.4.22
sentence-transformers>=2.3.0
# Relaxed LangChain/Graph versions to allow resolution
langchain>=0.1.10
langchain-community>=0.0.25
langchain-core>=0.1.30
langchain-google-genai>=0.0.9
langgraph>=0.0.24
langsmith
